Fixing namespacing, tests and documentation
* Calling postcode related methods directly from the Iceland module, for example `Iceland.all_postal_codes` has been depreciated. Use the `Iceland::PostalCodes` class instead.
* The kennitala class has been replaced with the more appropriately namespaced `Iceland::Kennitala` class. There is no deprecation involved.
* Updating all development dependencies to their newest version.
* Removing overcommit as it is fucking annoying
* Splitting specs into different files by functionality.
* Updating specs and documentation to reflect the depreciation of `Fixnum` in ruby 2.4.
