+++ .rspec
@@ -1,2 +1,3 @@
 --format documentation
 --color
+--deprecation-out /dev/null
