+++ .rubocop.yml
@@ -2,3 +2,5 @@ Metrics/AbcSize:
   Max: 20
 Metrics/ClassLength:
   Max: 150
+Metrics/BlockLength:
+  Max: 100
