+++ .travis.yml
@@ -1,6 +1,6 @@
 language: ruby
 rvm:
-  - 2.2.3
-before_install: gem install bundler -v 1.11.2
+  - 2.4.1
+before_install: gem install bundler -v 1.15
 notifications:
   email: false
