+++ lib/iceland.rb
@@ -1,7 +1,13 @@
+# frozen_string_literal: true
+
 require 'date'
 require 'yaml'
+require 'byebug'
 
 require 'iceland/version'
-require 'iceland/postal_codes'
-require 'iceland/kennitala'
-require 'iceland/kennitala_string'
+
+require 'iceland/iceland_postal_code'
+require 'iceland/iceland_kennitala'
+require 'iceland/iceland_kennitala_string'
+
+require 'iceland/depricated'
