+++ lib/iceland/iceland_kennitala_string.rb
@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 # Monkey patch for the String class
 class String
   # Converts a String to a Kennitala object
   #
   # @return [Kennitala]
   def to_kt
-    Kennitala.new(self)
+    Iceland::Kennitala.new(self)
   end
 end
