+++ lib/iceland/version.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Iceland
-  VERSION = '0.1.13'.freeze
+  VERSION = '0.2.0'
 end
