+++ spec/spec_helper.rb
@@ -1,5 +1,7 @@
-require 'codeclimate-test-reporter'
-CodeClimate::TestReporter.start
+# frozen_string_literal: true
+
+require 'simplecov'
+SimpleCov.start
 
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 require 'iceland'
